Allow typeName to be specified for table options, and raw values switch
Adding the typeName avoids a lot of problems where any two tables with the same field names but different data types will conflict with each other, and cause all sorts of unclear warnings when genertaing Gatsby nodes.

`typeName` is specified on a table-by-table basis, and if not present the standard ‘Airtable’ is used as before.

Raw values also cause a lot of warnings when the values of rows differ, so this commit also allows rawValues to be switched off globally, or on a table-by-table basis.