Add try catch for table link processing to catch invalid linked fields
This prevents fatal (and very hard to debug) errors from occurring when you are trying to add a table link for a field which is **not** a linked record tyoe in Airtable.

The catch block logs a warning to the user so they have a good clue about where to track down the problem in Airtable itself.